fix(gman): clamp the dice count before converting it to int
diceCountFor computed ceil(L / sqrt(ShadingRate)) and converted the
result to RtInt before clamping to [1, 16]. L is a finite raster
distance that can still exceed what an int holds, so the conversion
was undefined behaviour: macOS's cast saturates, masking it locally,
but the ubuntu/clang/asan+ubsan CI job aborted on it in
triangulation. Clamp in double, then convert, matching the settled
rule of clamp(ceil(L / sqrt(rate)), 1, 16) in real arithmetic. The
non-finite and non-positive fallbacks are unchanged.

tests/triangulation_test.cpp carries its own independent oracle for
the same formula, with the identical conversion-before-clamp defect;
under the full sanitizer build it aborted one call further in, once
the production path stopped aborting first. Fix it the same way.

Add a case to polygondicecount_test.cpp: a triangle whose longest
edge projects past INT_MAX pixels but is still finite, dicing to
256. Restoring the old conversion in either file makes its own test
abort under UBSan; both were reverted after confirming that and the
sources touched.

+++ tests/triangulation_test.cpp
@@ -143,14 +143,17 @@ int expectedDiceN(const GMANPoint& p0, const GMANPoint& p1, const GMANPoint& p2,
   if (!(shadingRate > (RtFloat)0.0) || !std::isfinite(L)) {
     return 16;
   }
-  int n = (int)std::ceil(L / std::sqrt((double)shadingRate));
-  if (n < 1) {
-    n = 1;
+  // Clamped in double before the RtInt conversion: mirrors
+  // gmanpatchpolyobjectmanager.cpp's own diceCountFor, since L can be
+  // finite and still past what an int can hold.
+  double n = std::ceil(L / std::sqrt((double)shadingRate));
+  if (n < 1.0) {
+    n = 1.0;
   }
-  if (n > 16) {
-    n = 16;
+  if (n > 16.0) {
+    n = 16.0;
   }
-  return n;
+  return (int)n;
 }
 
 // Rodrigues' rotation formula in double precision -- an independent
